+++ src/Body/BodyInterface.php
@@ -7,6 +7,8 @@ interface BodyInterface
 
     public function getContent(): ?string;
 
+    public function toArray(): ?array;
+
     /**
      * @return resource|null
      */