+++ src/Context/RequestContext.php
@@ -91,6 +91,20 @@ final class RequestContext implements RequestContextInterface, BodyInterface
         }
     }
 
+    public function toArray(): ?array
+    {
+        $requestBody = $this->getRequestBody();
+        if ($requestBody === null) {
+            return null;
+        }
+
+        try {
+            return $requestBody->toArray();
+        } catch (\Throwable $ex) {
+            return null;
+        }
+    }
+
     /**
      * @return resource|null
      */