+++ src/Context/ResponseContextInterface.php
@@ -15,6 +15,8 @@ interface ResponseContextInterface
 
     public function getContent(): ?string;
 
+    public function toArray(): ?array;
+
     /**
      * @return resource|null
      */