+++ tests/NativeHttpClientTest.php
@@ -60,6 +60,10 @@ class NativeHttpClientTest extends TestCase
                     'abc' => 'def',
                     'REQUEST_METHOD' => 'POST',
                 ],
+                'response-content-toarray' => [
+                    'abc' => 'def',
+                    'REQUEST_METHOD' => 'POST',
+                ],
             ],
         ];
         $this->assertSameResponseContentLog($expected, $logger->logs);
@@ -101,6 +105,10 @@ class NativeHttpClientTest extends TestCase
                     'abc' => 'def',
                     'REQUEST_METHOD' => 'POST',
                 ],
+                'response-content-toarray' => [
+                    'abc' => 'def',
+                    'REQUEST_METHOD' => 'POST',
+                ],
             ],
         ];
         $this->assertSameResponseContentLog($expected, $logger->logs);
@@ -134,7 +142,13 @@ class NativeHttpClientTest extends TestCase
             [
                 'message' => 'Response content: "200 http://127.0.0.1:8057/post"',
                 'request-content' => '{"abc":"def"}',
+                'request-content-json' => ['abc' => 'def'],
+                'request-content-toarray' => ['abc' => 'def'],
                 'response-content-json' => [
+                    '{"abc":"def"}' => '',
+                    'REQUEST_METHOD' => 'POST',
+                ],
+                'response-content-toarray' => [
                     '{"abc":"def"}' => '',
                     'REQUEST_METHOD' => 'POST',
                 ],
@@ -173,6 +187,10 @@ class NativeHttpClientTest extends TestCase
                 'message' => 'Response content: "200 http://127.0.0.1:8057/post"',
                 'request-content' => 'abc=def',
                 'response-content-json' => [
+                    'abc' => 'def',
+                    'REQUEST_METHOD' => 'POST',
+                ],
+                'response-content-toarray' => [
                     'abc' => 'def',
                     'REQUEST_METHOD' => 'POST',
                 ],
@@ -225,6 +243,10 @@ class NativeHttpClientTest extends TestCase
                     'abc' => 'def',
                     'REQUEST_METHOD' => 'POST',
                 ],
+                'response-content-toarray' => [
+                    'abc' => 'def',
+                    'REQUEST_METHOD' => 'POST',
+                ],
             ],
         ];
         $this->assertSameResponseContentLog($expected, $logger->logs);
@@ -267,6 +289,10 @@ class NativeHttpClientTest extends TestCase
                     'abc' => 'def',
                     'REQUEST_METHOD' => 'POST',
                 ],
+                'response-content-toarray' => [
+                    'abc' => 'def',
+                    'REQUEST_METHOD' => 'POST',
+                ],
             ],
         ];
         $this->assertSameResponseContentLog($expected, $logger->logs);
@@ -299,6 +325,10 @@ class NativeHttpClientTest extends TestCase
                 'message' => 'Response content: "200 http://127.0.0.1:8057/post"',
                 'request-content' => 'abc=def',
                 'response-content-json' => [
+                    'abc' => 'def',
+                    'REQUEST_METHOD' => 'POST',
+                ],
+                'response-content-toarray' => [
                     'abc' => 'def',
                     'REQUEST_METHOD' => 'POST',
                 ],
@@ -406,6 +436,10 @@ class NativeHttpClientTest extends TestCase
                     'abc' => 'def',
                     'REQUEST_METHOD' => 'POST',
                 ],
+                'response-content-toarray' => [
+                    'abc' => 'def',
+                    'REQUEST_METHOD' => 'POST',
+                ],
                 'response-headers-content-type' => ['application/json'],
                 'info-canceled' => false,
             ],
@@ -498,6 +532,19 @@ class NativeHttpClientTest extends TestCase
                     'HTTP_USER_AGENT' => 'Symfony HttpClient/Native',
                     'HTTP_HOST' => '127.0.0.1:8057',
                 ],
+                'response-content-toarray' => [
+                    'SERVER_PROTOCOL' => 'HTTP/1.1',
+                    'SERVER_NAME' => '127.0.0.1',
+                    'REQUEST_URI' => '/404',
+                    'REQUEST_METHOD' => 'POST',
+                    'HTTP_CONNECTION' => 'close',
+                    'HTTP_ACCEPT' => '*/*',
+                    'HTTP_CONTENT_LENGTH' => '7',
+                    'HTTP_CONTENT_TYPE' => 'application/x-www-form-urlencoded',
+                    'HTTP_ACCEPT_ENCODING' => 'gzip',
+                    'HTTP_USER_AGENT' => 'Symfony HttpClient/Native',
+                    'HTTP_HOST' => '127.0.0.1:8057',
+                ],
             ],
         ];
         $this->assertSameResponseContentLog($expected, $logger->logs);
@@ -537,6 +584,19 @@ class NativeHttpClientTest extends TestCase
                 'message' => 'Response content: "404 http://127.0.0.1:8057/404"',
                 'request-content' => 'abc=def',
                 'response-content-json' => [
+                    'SERVER_PROTOCOL' => 'HTTP/1.1',
+                    'SERVER_NAME' => '127.0.0.1',
+                    'REQUEST_URI' => '/404',
+                    'REQUEST_METHOD' => 'POST',
+                    'HTTP_CONNECTION' => 'close',
+                    'HTTP_ACCEPT' => '*/*',
+                    'HTTP_CONTENT_LENGTH' => '7',
+                    'HTTP_CONTENT_TYPE' => 'application/x-www-form-urlencoded',
+                    'HTTP_ACCEPT_ENCODING' => 'gzip',
+                    'HTTP_USER_AGENT' => 'Symfony HttpClient/Native',
+                    'HTTP_HOST' => '127.0.0.1:8057',
+                ],
+                'response-content-toarray' => [
                     'SERVER_PROTOCOL' => 'HTTP/1.1',
                     'SERVER_NAME' => '127.0.0.1',
                     'REQUEST_URI' => '/404',